+++ gnssrefl/rinex2snr.py
@@ -124,6 +124,9 @@ def run_rinex2snr(station, year_list, doy_list, isnr, orbtype, rate,dec_rate,arc
 
     """
     #
+    if archive == 'special':
+        print('The special option needs to be updated for new UNAVCO/Earthscope protocols. Please contact')
+        print('Tim Dittmann or Kelly Enloe at Earthscope if your file download does not work.')
     # do not allow illegal skipit values
     if skipit < 1:
         skipit = 1